+++ src/weathergen/model/model.py
@@ -19,23 +19,17 @@ import astropy_healpix.healpy
 import numpy as np
 import torch
 import torch.nn as nn
-from astropy_healpix import healpy
 from torch.utils.checkpoint import checkpoint
 
 from weathergen.common.config import Config
+from weathergen.model.encoder import EncoderModule
 from weathergen.model.engines import (
-    EmbeddingEngine,
     EnsPredictionHead,
     ForecastingEngine,
-    GlobalAssimilationEngine,
-    Local2GlobalAssimilationEngine,
-    LocalAssimilationEngine,
-    QueryAggregationEngine,
     TargetPredictionEngine,
     TargetPredictionEngineClassic,
 )
 from weathergen.model.layers import MLP, NamedLinear
-from weathergen.model.parametrised_prob_dist import LatentInterpolator
 from weathergen.model.utils import get_num_parameters
 from weathergen.utils.distributed import is_root
 from weathergen.utils.utils import get_dtype
@@ -269,14 +263,8 @@ class Model(torch.nn.Module):
         self.targets_num_channels = targets_num_channels
         self.targets_coords_size = targets_coords_size
 
-        self.ae_aggregation_engine: QueryAggregationEngine | None = None
-        self.ae_global_engine: GlobalAssimilationEngine | None = None
-        self.ae_local_engine: LocalAssimilationEngine | None = None
-        self.ae_local_global_engine: Local2GlobalAssimilationEngine | None = None
-        self.embed_engine: EmbeddingEngine | None = None
         self.embed_target_coords = None
         self.forecast_engine: ForecastingEngine | None = None
-        self.interpolate_latents: LatentInterpolator | None = None
         self.pred_adapter_kv = None
         self.pred_heads = None
         self.q_cells: torch.Tensor | None = None
@@ -288,62 +276,9 @@ class Model(torch.nn.Module):
         """Create each individual module of the model"""
         cf = self.cf
 
-        # determine stream names once so downstream components use consistent keys
-        self.stream_names = [str(stream_cfg["name"]) for stream_cfg in cf.streams]
-        # separate embedding networks for differnt observation types
-        self.embed_engine = EmbeddingEngine(cf, self.sources_size, self.stream_names)
-
-        ##############
-        # local assimilation engine
-        self.ae_local_engine = LocalAssimilationEngine(cf)
-
-        if cf.latent_noise_kl_weight > 0.0:
-            self.interpolate_latents = LatentInterpolator(
-                gamma=cf.latent_noise_gamma,
-                dim=cf.ae_local_dim_embed,
-                use_additive_noise=cf.latent_noise_use_additive_noise,
-                deterministic=cf.latent_noise_deterministic_latents,
-            )
-
-        ##############
-        # local -> global assimilation engine adapter
-        self.ae_local_global_engine = Local2GlobalAssimilationEngine(cf)
-
-        ##############
-        # learnable queries
-        if cf.ae_local_queries_per_cell:
-            s = (self.num_healpix_cells, cf.ae_local_num_queries, cf.ae_global_dim_embed)
-            q_cells = torch.rand(s, requires_grad=True) / cf.ae_global_dim_embed
-            # add meta data
-            q_cells[:, :, -8:-6] = (
-                (torch.arange(self.num_healpix_cells) / self.num_healpix_cells)
-                .unsqueeze(1)
-                .unsqueeze(1)
-                .repeat((1, cf.ae_local_num_queries, 2))
-            )
-            theta, phi = healpy.pix2ang(
-                nside=2**self.healpix_level, ipix=torch.arange(self.num_healpix_cells)
-            )
-            q_cells[:, :, -6:-3] = (
-                torch.cos(theta).unsqueeze(1).unsqueeze(1).repeat((1, cf.ae_local_num_queries, 3))
-            )
-            q_cells[:, :, -3:] = (
-                torch.sin(phi).unsqueeze(1).unsqueeze(1).repeat((1, cf.ae_local_num_queries, 3))
-            )
-            q_cells[:, :, -9] = torch.arange(cf.ae_local_num_queries)
-            q_cells[:, :, -10] = torch.arange(cf.ae_local_num_queries)
-        else:
-            s = (1, cf.ae_local_num_queries, cf.ae_global_dim_embed)
-            q_cells = torch.rand(s, requires_grad=True) / cf.ae_global_dim_embed
-        self.q_cells = torch.nn.Parameter(q_cells, requires_grad=True)
-
-        ##############
-        # query aggregation engine
-        self.ae_aggregation_engine = QueryAggregationEngine(cf, self.num_healpix_cells)
-
-        ##############
-        # global assimilation engine
-        self.ae_global_engine = GlobalAssimilationEngine(cf, self.num_healpix_cells)
+        self.encoder = EncoderModule(
+            cf, self.sources_size, self.targets_num_channels, self.targets_coords_size
+        )
 
         ###############
         # forecasting engine
@@ -365,6 +300,9 @@ class Model(torch.nn.Module):
         self.target_token_engines = torch.nn.ModuleDict()
         self.pred_adapter_kv = torch.nn.ModuleDict()
         self.pred_heads = torch.nn.ModuleDict()
+
+        # determine stream names once so downstream components use consistent keys
+        self.stream_names = [str(stream_cfg["name"]) for stream_cfg in cf.streams]
 
         for i_obs, si in enumerate(cf.streams):
             stream_name = self.stream_names[i_obs]
@@ -494,16 +432,19 @@ class Model(torch.nn.Module):
 
         cf = self.cf
         num_params_embed = [
-            get_num_parameters(self.embed_engine.embeds[name]) for name in self.stream_names
+            get_num_parameters(self.encoder.embed_engine.embeds[name]) for name in self.stream_names
         ]
         num_params_total = get_num_parameters(self)
-        num_params_ae_local = get_num_parameters(self.ae_local_engine.ae_local_blocks)
-        num_params_ae_global = get_num_parameters(self.ae_global_engine.ae_global_blocks)
+        num_params_ae_local = get_num_parameters(self.encoder.ae_local_engine.ae_local_blocks)
+        num_params_ae_global = get_num_parameters(self.encoder.ae_global_engine.ae_global_blocks)
 
-        num_params_q_cells = np.prod(self.q_cells.shape) if self.q_cells.requires_grad else 0
-        num_params_ae_adapater = get_num_parameters(self.ae_local_global_engine.ae_adapter)
+        num_params_q_cells = (
+            np.prod(self.encoder.q_cells.shape) if self.encoder.q_cells.requires_grad else 0
+        )
+        num_params_ae_adapater = get_num_parameters(self.encoder.ae_local_global_engine.ae_adapter)
+
         num_params_ae_aggregation = get_num_parameters(
-            self.ae_aggregation_engine.ae_aggregation_blocks
+            self.encoder.ae_aggregation_engine.ae_aggregation_blocks
         )
 
         num_params_fe = get_num_parameters(self.forecast_engine.fe_blocks)
@@ -559,11 +500,16 @@ class Model(torch.nn.Module):
             new_params : Dictionary with (renamed) model parameters
         """
         params_cleanup = {
-            "embeds": "embed_engine.embeds",  # EmbeddingEngine
-            "ae_local_blocks": "ae_local_engine.ae_local_blocks",  # LocalAssimilationEngine
-            "ae_adapter": "ae_local_global_engine.ae_adapter",  # Local2GlobalAssimilationEngine
-            "ae_global_blocks": "ae_global_engine.ae_global_blocks",  # GlobalAssimilationEngine
-            "fe_blocks": "forecast_engine.fe_blocks",  # ForecastingEngine
+            # EmbeddingEngine
+            "embeds": "encoder.embed_engine.embeds",
+            # LocalAssimilationEngine
+            "ae_local_blocks": "encoder.ae_local_engine.ae_local_blocks",
+            # Local2GlobalAssimilationEngine
+            "ae_adapter": "encoder.ae_local_global_engine.ae_adapter",
+            # GlobalAssimilationEngine
+            "ae_global_blocks": "encoder.ae_global_engine.ae_global_blocks",
+            # ForecastingEngine
+            "fe_blocks": "forecast_engine.fe_blocks",
         }
 
         new_params = {}
@@ -606,13 +552,7 @@ class Model(torch.nn.Module):
             A list containing all prediction results
         """
 
-        # embed
-        tokens = self.embed_cells(model_params, sample)
-
-        # local assimilation engine and adapter
-        tokens, posteriors = self.assimilate_local(model_params, tokens, sample)
-
-        tokens = self.assimilate_global(model_params, tokens)
+        tokens, posteriors = self.encoder(model_params, sample)
 
         # roll-out in latent space
         preds_all = []
@@ -650,171 +590,6 @@ class Model(torch.nn.Module):
         latents["posteriors"] = posteriors
 
         return ModelOutput(physical=preds_all, latent=latents)
-
-    #########################################
-    def embed_cells(self, model_params: ModelParams, sample) -> torch.Tensor:
-        """Embeds input data for each stream separately and rearranges it to cell-wise order
-        Args:
-            model_params : Query and embedding parameters
-            streams_data : Used to initialize first tokens for pre-processing
-        Returns:uv
-            Tokens for local assimilation
-        """
-
-        device = next(self.parameters()).device
-        tokens_all = self.embed_engine(sample, model_params.pe_embed, self.dtype, device)
-
-        return tokens_all
-
-    #########################################
-    def assimilate_local(
-        self, model_params: ModelParams, tokens: torch.Tensor, sample: torch.Tensor
-    ) -> torch.Tensor:
-        """Processes embedded tokens locally and prepares them for the global assimilation
-        Args:
-            model_params : Query and embedding parameters
-            tokens : Input tokens to be processed by local assimilation
-            cell_lens : Used to identify range of tokens to use from generated tokens in cell
-                embedding
-        Returns:
-            Tokens for global assimilation
-        """
-
-        cell_lens = sample.source_cell_lens
-        batch_size = (
-            self.cf.batch_size_per_gpu if self.training else self.cf.batch_size_validation_per_gpu
-        )
-
-        s = self.q_cells.shape
-        # print( f'{np.prod(np.array(tokens.shape))} :: {np.prod(np.array(s))}'
-        #        + ':: {np.prod(np.array(tokens.shape))/np.prod(np.array(s))}')
-        # TODO: test if positional encoding is needed here
-        if self.cf.ae_local_queries_per_cell:
-            tokens_global = (self.q_cells + model_params.pe_global).repeat(batch_size, 1, 1)
-        else:
-            tokens_global = (
-                self.q_cells.repeat(self.num_healpix_cells, 1, 1) + model_params.pe_global
-            )
-        q_cells_lens = torch.cat(
-            [model_params.q_cells_lens[0].unsqueeze(0)]
-            + [model_params.q_cells_lens[1:] for _ in range(batch_size)]
-        )
-
-        # local assimilation model
-        # for block in self.ae_local_blocks:
-        #     tokens = checkpoint(block, tokens, cell_lens, use_reentrant=False)
-
-        # if self.cf.latent_noise_kl_weight > 0.0:
-        #     tokens, posteriors = self.interpolate_latents.interpolate_with_noise(
-        #         tokens, sampling=self.training
-        #     )
-        # else:
-        #     tokens, posteriors = tokens, 0.0
-
-        # for block in self.ae_adapter:
-        #     tokens_global = checkpoint(
-        #         block,
-        #         tokens_global,
-        #         tokens,
-        #         q_cells_lens,
-        #         cell_lens,
-        #         use_reentrant=False,
-        #     )
-
-        # work around to bug in flash attention for hl>=5
-
-        istep = 0
-
-        cell_lens = cell_lens[istep][1:]
-        clen = self.num_healpix_cells // (2 if self.cf.healpix_level <= 5 else 8)
-        tokens_global_unmasked_all = []
-        posteriors = []
-        zero_pad = torch.zeros(1, device=tokens.device, dtype=torch.int32)
-        for i in range((cell_lens.shape[0]) // clen):
-            # make sure we properly catch all elements in last chunk
-            i_end = (i + 1) * clen if i < (cell_lens.shape[0] // clen) - 1 else cell_lens.shape[0]
-            l0, l1 = (
-                (0 if i == 0 else cell_lens[: i * clen].cumsum(0)[-1]),
-                cell_lens[:i_end].cumsum(0)[-1],
-            )
-
-            tokens_c = tokens[l0:l1]
-            tokens_global_c = tokens_global[i * clen : i_end]
-            cell_lens_c = torch.cat([zero_pad, cell_lens[i * clen : i_end]])
-            q_cells_lens_c = q_cells_lens[: cell_lens_c.shape[0]]
-
-            # local assimilation model
-            tokens_c = self.ae_local_engine(tokens_c, cell_lens_c, use_reentrant=False)
-
-            if self.cf.latent_noise_kl_weight > 0.0:
-                tokens_c, posteriors_c = self.interpolate_latents.interpolate_with_noise(
-                    tokens_c, sampling=self.training
-                )
-                posteriors += [posteriors_c]
-            else:
-                tokens_c, posteriors = tokens_c, 0.0
-
-            # create mask for global tokens, without first element (used for padding)
-            mask_c = cell_lens_c[1:].to(torch.bool)
-            tokens_global_unmasked_c = tokens_global_c[mask_c]
-            q_cells_lens_unmasked_c = torch.cat([zero_pad, q_cells_lens_c[1:][mask_c]])
-            cell_lens_unmasked_c = torch.cat([zero_pad, cell_lens_c[1:][mask_c]])
-
-            if l0 == l1 or tokens_c.shape[0] == 0:
-                tokens_global_unmasked_all += [tokens_global_unmasked_c]
-                continue
-
-            # local to global adapter engine
-            tokens_global_unmasked_c = self.ae_local_global_engine(
-                tokens_c,
-                tokens_global_unmasked_c,
-                q_cells_lens_unmasked_c,
-                cell_lens_unmasked_c,
-                use_reentrant=False,
-            )
-
-            tokens_global_unmasked_all += [tokens_global_unmasked_c]
-
-        tokens_global_unmasked = torch.cat(tokens_global_unmasked_all)
-
-        # query aggregation engine on the query tokens in unmasked cells
-        # (applying this here assumes batch_size=1)
-        # permute to use ae_local_num_queries as the batchsize and no_of_tokens
-        # as seq len for flash attention
-        tokens_global_unmasked = torch.permute(tokens_global_unmasked, [1, 0, 2])
-        tokens_global_unmasked = self.ae_aggregation_engine(
-            tokens_global_unmasked, use_reentrant=False
-        )
-        tokens_global_unmasked = torch.permute(tokens_global_unmasked, [1, 0, 2])
-
-        # create mask from cell lens
-        mask = cell_lens.to(torch.bool)
-
-        # fill empty tensor using mask for positions of unmasked tokens
-        tokens_global[mask] = tokens_global_unmasked.to(tokens_global.dtype)
-
-        # recover batch dimension and build global token list
-        tokens_global = (
-            tokens_global.reshape([batch_size, self.num_healpix_cells, s[-2], s[-1]])
-            + model_params.pe_global
-        ).flatten(1, 2)
-
-        return tokens_global, posteriors
-
-    #########################################
-    def assimilate_global(self, model_params: ModelParams, tokens: torch.Tensor) -> torch.Tensor:
-        """Performs transformer based global assimilation in latent space
-        Args:
-            model_params : Query and embedding parameters (never used)
-            tokens : Input tokens to be pre-processed by global assimilation
-        Returns:
-            Latent representation of the model
-        """
-
-        # global assimilation engine and adapter
-        tokens = self.ae_global_engine(tokens, use_reentrant=False)
-
-        return tokens
 
     #########################################
     def forecast(self, model_params: ModelParams, tokens: torch.Tensor, fstep: int) -> torch.Tensor:
